+++ apps/selvbetjening-ui/src/inntektsjustering/1-innledning/InntektsjusteringInnledning.tsx
@@ -18,7 +18,6 @@ import { SpraakVelger } from '../../common/spraak/SpraakVelger.tsx'
 import { HarIkkeOMSSakIGjenny } from './HarIkkeOMSSakIGjenny.tsx'
 import { FeilVedSjekkAvOMSSakIGjenny } from './FeilVedSjekkAvOMSSakIGjenny.tsx'
 import { LogEvents, useAmplitude } from '../../hooks/useAmplitude.ts'
-import { inntektjusteringBasePath } from '../inntektjusteringRoutes.tsx'
 
 export const InntektsjusteringInnledning = () => {
     const navigate = useNavigate()
@@ -169,7 +168,7 @@ export const InntektsjusteringInnledning = () => {
                             <Button
                                 icon={<ArrowRightIcon aria-hidden />}
                                 iconPosition="right"
-                                onClick={() => navigate(`${inntektjusteringBasePath}/inntekt-til-neste-aar`)}
+                                onClick={() => navigate(`/inntekt-til-neste-aar`)}
                             >
                                 {innhold.startUtfyllingKnapp?.[spraak]}
                             </Button>
