+++ apps/selvbetjening-ui/src/inntektsjustering/2-inntekt-til-neste-aar/skjemaer/AttenTilSekstiEnAarSkjema.tsx
@@ -13,7 +13,6 @@ import { ControlledInntektTextField } from '../../../common/inntekt/ControlledIn
 import { Alder } from '../../../types/person.ts'
 import { SideLaster } from '../../../common/SideLaster.tsx'
 import { inntektSkjemaValuesTilInntekt, inntektTilInntektSkjemaValues } from '../../../utils/inntekt.ts'
-import { inntektjusteringBasePath } from '../../inntektjusteringRoutes.tsx'
 
 export const AttenTilSekstiEnAarSkjema = () => {
     const spraak = useSpraak()
@@ -48,7 +47,7 @@ export const AttenTilSekstiEnAarSkjema = () => {
 
     const onInntektSubmit = (inntekt: Inntekt) => {
         inntektDispatch.setInntekt(inntekt)
-        navigate(`${inntektjusteringBasePath}/oppsummering`)
+        navigate(`/oppsummering`)
     }
 
     const { hovedinnhold, arbeidsinntekt, naeringsinntekt, inntektFraUtland, sumAvInntekt } =
@@ -110,7 +109,7 @@ export const AttenTilSekstiEnAarSkjema = () => {
                     </SumAvOppgittInntekt>
 
                     <NavigasjonMeny
-                        tilbakePath={`${inntektjusteringBasePath}/innledning`}
+                        tilbakePath={`/innledning`}
                         onNeste={handleSubmit((inntekt) => onInntektSubmit(inntektSkjemaValuesTilInntekt(inntekt)))}
                     />
                 </VStack>
