+++ apps/selvbetjening-ui/src/inntektsjustering/2-inntekt-til-neste-aar/skjemaer/SekstiToTilSekstiSeksAarSkjema.tsx
@@ -16,7 +16,6 @@ import { ControlledMaanedVelger } from '../../../common/maanedVelger/ControlledM
 import { formaterFieldErrors } from '../../../utils/error.ts'
 import { SideLaster } from '../../../common/SideLaster.tsx'
 import { inntektSkjemaValuesTilInntekt, inntektTilInntektSkjemaValues } from '../../../utils/inntekt.ts'
-import { inntektjusteringBasePath } from '../../inntektjusteringRoutes.tsx'
 
 export const SekstiToTilSekstiSeksAarSkjema = ({ innloggetBruker }: { innloggetBruker: IInnloggetBruker }) => {
     const spraak = useSpraak()
@@ -54,7 +53,7 @@ export const SekstiToTilSekstiSeksAarSkjema = ({ innloggetBruker }: { innloggetB
 
     const onInntektSubmit = (inntekt: Inntekt) => {
         inntektDispatch.setInntekt(inntekt)
-        navigate(`${inntektjusteringBasePath}/oppsummering`)
+        navigate(`/oppsummering`)
     }
 
     const {
@@ -256,7 +255,7 @@ export const SekstiToTilSekstiSeksAarSkjema = ({ innloggetBruker }: { innloggetB
                     )}
 
                     <NavigasjonMeny
-                        tilbakePath={`${inntektjusteringBasePath}/innledning`}
+                        tilbakePath={`/innledning`}
                         onNeste={handleSubmit((inntekt) => onInntektSubmit(inntektSkjemaValuesTilInntekt(inntekt)))}
                     />
                 </VStack>
